Moved some code to commons project
