+++ src/main/java/gov/nasa/pds/harvest/dao/RegistryManager.java
@@ -9,8 +9,11 @@ import gov.nasa.pds.registry.common.es.client.EsClientFactory;
 import gov.nasa.pds.registry.common.es.dao.ProductDao;
 import gov.nasa.pds.registry.common.es.dao.dd.DataDictionaryDao;
 import gov.nasa.pds.registry.common.es.dao.schema.SchemaDao;
+import gov.nasa.pds.registry.common.es.service.MissingFieldsProcessor;
 import gov.nasa.pds.registry.common.es.service.ProductService;
+import gov.nasa.pds.registry.common.es.service.SchemaUpdater;
 import gov.nasa.pds.registry.common.meta.FieldNameCache;
+import gov.nasa.pds.registry.common.meta.MetadataNormalizer;
 import gov.nasa.pds.registry.common.util.CloseUtils;
 
 
@@ -23,6 +26,8 @@ public class RegistryManager
 {
     // Singleton
     private static RegistryManager instance = null;
+
+    private RegistryCfg cfg;
     
     // Elasticsearch client
     private RestClient esClient;
@@ -45,6 +50,7 @@ public class RegistryManager
      */
     private RegistryManager(RegistryCfg cfg) throws Exception
     {
+        this.cfg = cfg;
         if(cfg.url == null || cfg.url.isEmpty()) throw new IllegalArgumentException("Missing Registry URL");
         
         esClient = EsClientFactory.createRestClient(cfg.url, cfg.authFile);
@@ -161,6 +167,28 @@ public class RegistryManager
     public FieldNameCache getFieldNameCache()
     {
         return fieldNameCache;
+    }
+
+    
+    /**
+     * Create new missing field processor
+     * @return new missing field processor object
+     * @throws Exception an exception
+     */
+    public MissingFieldsProcessor createMissingFieldsProcessor() throws Exception
+    {
+        SchemaUpdater su = new SchemaUpdater(cfg, ddDao, schemaDao);
+        return new MissingFieldsProcessor(su, fieldNameCache);
+    }
+
+
+    /**
+     * Create new metadata normalizer
+     * @return new metadata normalizer object
+     */
+    public MetadataNormalizer createMetadataNormalizer()
+    {
+        return new MetadataNormalizer(fieldNameCache);
     }
 
 }
