+++ src/main/java/gov/nasa/pds/harvest/mq/CollectionInventoryConsumer.java
@@ -6,8 +6,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import gov.nasa.pds.harvest.mq.msg.CollectionInventoryMessage;
-import gov.nasa.pds.harvest.proc.CollectionInventoryProcessor;
 import gov.nasa.pds.registry.common.cfg.RegistryCfg;
+import gov.nasa.pds.registry.common.es.service.CollectionInventoryWriter;
 import gov.nasa.pds.registry.common.util.ExceptionUtils;
 
 /**
@@ -17,7 +17,7 @@ import gov.nasa.pds.registry.common.util.ExceptionUtils;
 public class CollectionInventoryConsumer
 {
     protected Logger log;
-    private CollectionInventoryProcessor proc;
+    private CollectionInventoryWriter proc;
     
     
     /**
@@ -28,7 +28,7 @@ public class CollectionInventoryConsumer
     public CollectionInventoryConsumer(RegistryCfg cfg) throws Exception
     {
         log = LogManager.getLogger(this.getClass());
-        proc = new CollectionInventoryProcessor(cfg);
+        proc = new CollectionInventoryWriter(cfg);
     }
