+++ src/main/java/gov/nasa/pds/harvest/mq/ProductConsumer.java
@@ -7,19 +7,15 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import gov.nasa.pds.harvest.cfg.HarvestCfg;
-import gov.nasa.pds.harvest.dao.RegistryManager;
 import gov.nasa.pds.harvest.dao.RegistryService;
 import gov.nasa.pds.harvest.job.Job;
 import gov.nasa.pds.harvest.job.JobFactory;
 import gov.nasa.pds.harvest.mq.msg.ProductMessage;
 import gov.nasa.pds.harvest.proc.ProductProcessor;
-import gov.nasa.pds.harvest.util.out.RegistryDocWriter;
 import gov.nasa.pds.registry.common.cfg.RegistryCfg;
 import gov.nasa.pds.registry.common.es.dao.DataLoader;
-import gov.nasa.pds.registry.common.es.dao.dd.DataDictionaryDao;
-import gov.nasa.pds.registry.common.es.dao.schema.SchemaDao;
-import gov.nasa.pds.registry.common.es.service.SchemaUpdater;
 import gov.nasa.pds.registry.common.util.ExceptionUtils;
+import gov.nasa.pds.registry.common.util.doc.RegistryDocWriter;
 
 
 /**
@@ -35,7 +31,6 @@ public class ProductConsumer
     private RegistryDocWriter registryDocWriter;
     private ProductProcessor proc;
     private DataLoader dataLoader;
-    private SchemaUpdater schemaUpdater;
 
     
     /**
@@ -54,11 +49,6 @@ public class ProductConsumer
         proc = new ProductProcessor(harvestCfg, registryDocWriter);
         
         dataLoader = new DataLoader(registryCfg.url, registryCfg.indexName, registryCfg.authFile);
-        
-        SchemaDao schemaDao = RegistryManager.getInstance().getSchemaDao();
-        DataDictionaryDao ddDao = RegistryManager.getInstance().getDataDictionaryDao();
-
-        schemaUpdater = new SchemaUpdater(registryCfg, ddDao, schemaDao);
     }
     
     
@@ -118,32 +108,7 @@ public class ProductConsumer
                 // Ignore this file
             }
         }
-        
-        // Update Elasticsearch schema if needed
-        if(!registryDocWriter.getMissingFields().isEmpty())
-        {
-            try
-            {
-                // Update schema
-                schemaUpdater.updateSchema(registryDocWriter.getMissingFields(), registryDocWriter.getMissingXsds());
-                // Update cache
-                RegistryManager.getInstance().getFieldNameCache().update();
-            }
-            catch(Exception ex)
-            {
-                log.error("Could not update Elasticsearch schema. " + ExceptionUtils.getMessage(ex));
-
-                // Ignore the whole batch for now
-                log.error("Following files will not be processed:");
-                for(String file: filesToProcess)
-                {
-                    log.error(file);
-                }
                 
-                return true;
-            }
-        }
-        
         // Load the data into Elasticsearch
         try
         {
